+++ MovieAPI.Infrastructure/Extensions/DictionaryExtension.cs
@@ -5,7 +5,7 @@ namespace MovieAPI.Infrastructure.Extensions
 {
     public static class DictionaryExtension
     {
-        public static NameValueCollection ConvertToQueryParameters(this Dictionary<string, string?> dictionaryParameters)
+        public static NameValueCollection ConvertToQueryParameters(this Dictionary<string, string> dictionaryParameters)
         {
             var parameters = HttpUtility.ParseQueryString(string.Empty);
             foreach (var pair in dictionaryParameters)
@@ -13,7 +13,7 @@ namespace MovieAPI.Infrastructure.Extensions
                 parameters.Add(pair.Key, pair.Value);
             }
 
-            return parameters;
+            return parameters ?? new NameValueCollection();
         }
     }
 }
